requirements.txt
rasterio
numpy
matplotlib
scikit-learn
requests
pillow
